ATR-256: refactoring; add MemoryDiagnoser
