+++ src/Acuminator/Acuminator.Benchmark/Acuminator.Benchmark.csproj
@@ -227,7 +227,8 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RoslynCodeAnalysis.cs" />
+    <Compile Include="RoslynCodeAnalysisBenchmark.cs" />
+    <Compile Include="RoslynHelper.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
